Better logic behind storing logfiles
now for every game that is being logged a map-specific directory
is created. Upon creation a file named "borders.csv" is added which
contains a representation of the graph edges.
With logging active two files with the current datetimes plus a
character "b" or "t" as names are created per game. These csv-files
contain the state after each processed command as rows, the state
of troops and the state of belonging to a player respectively.