Proper Implementation of game history visualization via colored graphs
visualisierungSpiel.py now lets the user interactively choose a
gamelog csv-file. Then it produces a graph based on the map the
gamelog is associated with, namely the borders.csv file in the same
directory.
Interact with the graph!
Move vertices: Use your mouse to pull the vertices around
press "c": randomize the colors representing the different players
press "n"/"l": go to the next/previous state of the game